+++ crates/gauntlet/src/lib.rs
@@ -40,6 +40,7 @@ use report::Status;
 use report::UnmatchedStatus;
 pub use repository::Repository;
 use wdl::analysis::Analyzer;
+use wdl::analysis::DiagnosticsConfig;
 use wdl::analysis::rules;
 use wdl::ast::Diagnostic;
 use wdl::lint::LintVisitor;
@@ -174,7 +175,7 @@ pub async fn gauntlet(args: Args) -> Result<()> {
 
         let analyzer = Analyzer::new_with_validator(
             // Don't bother duplicating analysis warnings for arena mode
-            if args.arena { Vec::new() } else { rules() },
+            DiagnosticsConfig::new(if args.arena { Vec::new() } else { rules() }),
             move |_: (), _, _, _| async move {},
             move || {
                 let mut validator = if !args.arena {