+++ crates/wdl-engine/tests/inputs.rs
@@ -37,6 +37,7 @@ use pretty_assertions::StrComparison;
 use rayon::prelude::*;
 use wdl_analysis::AnalysisResult;
 use wdl_analysis::Analyzer;
+use wdl_analysis::DiagnosticsConfig;
 use wdl_analysis::rules;
 use wdl_analysis::types::Types;
 use wdl_ast::Diagnostic;
@@ -197,7 +198,7 @@ async fn main() {
     println!("\nrunning {} tests\n", tests.len());
 
     // Start with a single analysis pass over all the test files
-    let analyzer = Analyzer::new(rules(), |_, _, _, _| async {});
+    let analyzer = Analyzer::new(DiagnosticsConfig::new(rules()), |_, _, _, _| async {});
     for test in &tests {
         analyzer
             .add_directory(test.clone())