+++ crates/wdl-engine/tests/tasks.rs
@@ -47,6 +47,7 @@ use tempfile::TempDir;
 use walkdir::WalkDir;
 use wdl_analysis::AnalysisResult;
 use wdl_analysis::Analyzer;
+use wdl_analysis::DiagnosticsConfig;
 use wdl_analysis::document::Document;
 use wdl_analysis::rules;
 use wdl_ast::Diagnostic;
@@ -411,7 +412,7 @@ async fn main() {
     println!("\nrunning {} tests\n", tests.len());
 
     // Start with a single analysis pass over all the test files
-    let analyzer = Analyzer::new(rules(), |_, _, _, _| async {});
+    let analyzer = Analyzer::new(DiagnosticsConfig::new(rules()), |_, _, _, _| async {});
     for test in &tests {
         analyzer
             .add_directory(test.clone())